+++ CHANGELOG.md
@@ -15,15 +15,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.0.2] - 2025-08-31
  * create Result model
  * calculate LeaguePoints for Result
+ * League table, adding results, ordering
 
 ## WIP
  * None
 
 ## Todo
- * League table, adding results, ordering
  * Results from file
  * League table to file
 
 ## Maybes
  * also support stdin/stdout
- * move leaguepoints logic into leaguepoints class